feat: remove business logic from chat.tsx
