+++ app/components/chat/chat.tsx
@@ -4,23 +4,20 @@ import { ChatInput } from "@/app/components/chat-input/chat-input"
 import { Conversation } from "@/app/components/chat/conversation"
 import { useModel } from "@/app/components/chat/use-model"
 import { useChatDraft } from "@/app/hooks/use-chat-draft"
-import { toast } from "@/components/ui/toast"
-import { getOrCreateGuestUserId } from "@/lib/api"
 import { useChats } from "@/lib/chat-store/chats/provider"
 import { useMessages } from "@/lib/chat-store/messages/provider"
 import { useChatSession } from "@/lib/chat-store/session/provider"
-import { MESSAGE_MAX_LENGTH, SYSTEM_PROMPT_DEFAULT } from "@/lib/config"
-import { Attachment } from "@/lib/file-handling"
-import { API_ROUTE_CHAT } from "@/lib/routes"
 import { useUserPreferences } from "@/lib/user-preference-store/provider"
 import { useUser } from "@/lib/user-store/provider"
 import { cn } from "@/lib/utils"
-import { useChat } from "@ai-sdk/react"
 import { AnimatePresence, motion } from "motion/react"
 import dynamic from "next/dynamic"
-import { redirect, useSearchParams } from "next/navigation"
-import { Suspense, useCallback, useMemo, useRef, useState } from "react"
+import { redirect } from "next/navigation"
+import { Suspense, useMemo } from "react"
+import { useChatActions } from "./use-chat-actions"
 import { useChatHandlers } from "./use-chat-handlers"
+import { useChatInitialization } from "./use-chat-initialization"
+import { useChatState } from "./use-chat-state"
 import { useChatUtils } from "./use-chat-utils"
 import { useFileUpload } from "./use-file-upload"
 
@@ -33,21 +30,6 @@ const DialogAuth = dynamic(
   () => import("./dialog-auth").then((mod) => mod.DialogAuth),
   { ssr: false }
 )
-
-function SearchParamsProvider({
-  setInput,
-}: {
-  setInput: (input: string) => void
-}) {
-  const searchParams = useSearchParams()
-  const prompt = searchParams.get("prompt")
-
-  if (prompt && typeof window !== "undefined") {
-    requestAnimationFrame(() => setInput(prompt))
-  }
-
-  return null
-}
 
 export function Chat() {
   const { chatId } = useChatSession()
@@ -66,11 +48,42 @@ export function Chat() {
 
   const { messages: initialMessages, cacheAndAddMessage } = useMessages()
   const { user } = useUser()
-  const [isSubmitting, setIsSubmitting] = useState(false)
   const { preferences } = useUserPreferences()
-  const [hasDialogAuth, setHasDialogAuth] = useState(false)
-  const [enableSearch, setEnableSearch] = useState(false)
+  const { draftValue, clearDraft } = useChatDraft(chatId)
 
+  // Initialize chat and get useChat hooks
+  const {
+    messages,
+    input,
+    handleSubmit,
+    status,
+    reload,
+    stop,
+    setMessages,
+    setInput,
+    append,
+    isAuthenticated,
+    systemPrompt,
+    hasSentFirstMessageRef,
+  } = useChatInitialization({
+    initialMessages,
+    draftValue,
+    cacheAndAddMessage,
+    chatId,
+    user,
+  })
+
+  // Manage component state
+  const {
+    isSubmitting,
+    setIsSubmitting,
+    hasDialogAuth,
+    setHasDialogAuth,
+    enableSearch,
+    setEnableSearch,
+  } = useChatState({ setInput })
+
+  // File upload functionality
   const {
     files,
     setFiles,
@@ -81,6 +94,7 @@ export function Chat() {
     handleFileRemove,
   } = useFileUpload()
 
+  // Model selection
   const { selectedModel, handleModelChange } = useModel({
     currentChat: currentChat || null,
     user,
@@ -88,65 +102,7 @@ export function Chat() {
     chatId,
   })
 
-  const systemPrompt = useMemo(
-    () => user?.system_prompt || SYSTEM_PROMPT_DEFAULT,
-    [user?.system_prompt]
-  )
-
-  const hasSentFirstMessageRef = useRef(false)
-  const prevChatIdRef = useRef<string | null>(chatId)
-  const isAuthenticated = useMemo(() => !!user?.id, [user?.id])
-
-  const { draftValue, clearDraft } = useChatDraft(chatId)
-
-  // Handle errors directly in onError callback
-  const handleError = useCallback((error: Error) => {
-    console.error("Chat error:", error)
-    console.error("Error message:", error.message)
-    // The server now properly forwards error messages via getErrorMessage
-    // So we can use the error message directly
-    let errorMsg = error.message || "Something went wrong."
-
-    // If the error message is still generic, provide a fallback
-    if (errorMsg === "An error occurred" || errorMsg === "fetch failed") {
-      errorMsg = "Something went wrong. Please try again."
-    }
-
-    toast({
-      title: errorMsg,
-      status: "error",
-    })
-  }, [])
-
-  const {
-    messages,
-    input,
-    handleSubmit,
-    status,
-    error,
-    reload,
-    stop,
-    setMessages,
-    setInput,
-    append,
-  } = useChat({
-    api: API_ROUTE_CHAT,
-    initialMessages,
-    initialInput: draftValue,
-    onFinish: cacheAndAddMessage,
-    onError: handleError,
-  })
-
-  // Reset messages when navigating from a chat to home (not on every render)
-  if (
-    prevChatIdRef.current !== null &&
-    chatId === null &&
-    messages.length > 0
-  ) {
-    setMessages([])
-  }
-  prevChatIdRef.current = chatId
-
+  // Chat utilities
   const { checkLimitsAndNotify, ensureChatExists } = useChatUtils({
     isAuthenticated,
     chatId,
@@ -158,6 +114,7 @@ export function Chat() {
     setHasDialogAuth,
   })
 
+  // Message handlers
   const { handleInputChange, handleDelete, handleEdit } = useChatHandlers({
     messages,
     setMessages,
@@ -165,103 +122,8 @@ export function Chat() {
     chatId,
   })
 
-  const submit = useCallback(async () => {
-    setIsSubmitting(true)
-
-    const uid = await getOrCreateGuestUserId(user)
-    if (!uid) {
-      setIsSubmitting(false)
-      return
-    }
-
-    const optimisticId = `optimistic-${Date.now().toString()}`
-    const optimisticAttachments =
-      files.length > 0 ? createOptimisticAttachments(files) : []
-
-    const optimisticMessage = {
-      id: optimisticId,
-      content: input,
-      role: "user" as const,
-      createdAt: new Date(),
-      experimental_attachments:
-        optimisticAttachments.length > 0 ? optimisticAttachments : undefined,
-    }
-
-    setMessages((prev) => [...prev, optimisticMessage])
-    setInput("")
-
-    const submittedFiles = [...files]
-    setFiles([])
-
-    try {
-      const allowed = await checkLimitsAndNotify(uid)
-      if (!allowed) {
-        setMessages((prev) => prev.filter((m) => m.id !== optimisticId))
-        cleanupOptimisticAttachments(optimisticMessage.experimental_attachments)
-        return
-      }
-
-      const currentChatId = await ensureChatExists(uid)
-      if (!currentChatId) {
-        setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
-        cleanupOptimisticAttachments(optimisticMessage.experimental_attachments)
-        return
-      }
-
-      if (input.length > MESSAGE_MAX_LENGTH) {
-        toast({
-          title: `The message you submitted was too long, please submit something shorter. (Max ${MESSAGE_MAX_LENGTH} characters)`,
-          status: "error",
-        })
-        setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
-        cleanupOptimisticAttachments(optimisticMessage.experimental_attachments)
-        return
-      }
-
-      let attachments: Attachment[] | null = []
-      if (submittedFiles.length > 0) {
-        attachments = await handleFileUploads(uid, currentChatId)
-        if (attachments === null) {
-          setMessages((prev) => prev.filter((m) => m.id !== optimisticId))
-          cleanupOptimisticAttachments(
-            optimisticMessage.experimental_attachments
-          )
-          return
-        }
-      }
-
-      const options = {
-        body: {
-          chatId: currentChatId,
-          userId: uid,
-          model: selectedModel,
-          isAuthenticated,
-          systemPrompt: systemPrompt || SYSTEM_PROMPT_DEFAULT,
-          enableSearch,
-        },
-        experimental_attachments: attachments || undefined,
-      }
-
-      handleSubmit(undefined, options)
-      setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
-      cleanupOptimisticAttachments(optimisticMessage.experimental_attachments)
-      cacheAndAddMessage(optimisticMessage)
-      clearDraft()
-      hasSentFirstMessageRef.current = true
-
-      // Bump existing chats to top (non-blocking, after submit)
-      // If messages.length === 0, this is a new chat that was just created
-      if (messages.length > 0) {
-        bumpChat(currentChatId)
-      }
-    } catch (submitError) {
-      setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
-      cleanupOptimisticAttachments(optimisticMessage.experimental_attachments)
-      toast({ title: "Failed to send message", status: "error" })
-    } finally {
-      setIsSubmitting(false)
-    }
-  }, [
+  // Chat actions
+  const { submit, handleSuggestion, handleReload } = useChatActions({
     user,
     files,
     createOptimisticAttachments,
@@ -276,101 +138,17 @@ export function Chat() {
     selectedModel,
     isAuthenticated,
     systemPrompt,
+    enableSearch,
     handleSubmit,
+    append,
     cacheAndAddMessage,
     clearDraft,
-    messages.length,
+    messages,
     bumpChat,
-  ])
-
-  const handleSuggestion = useCallback(
-    async (suggestion: string) => {
-      setIsSubmitting(true)
-      const optimisticId = `optimistic-${Date.now().toString()}`
-      const optimisticMessage = {
-        id: optimisticId,
-        content: suggestion,
-        role: "user" as const,
-        createdAt: new Date(),
-      }
-
-      setMessages((prev) => [...prev, optimisticMessage])
-
-      try {
-        const uid = await getOrCreateGuestUserId(user)
-
-        if (!uid) {
-          setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
-          return
-        }
-
-        const allowed = await checkLimitsAndNotify(uid)
-        if (!allowed) {
-          setMessages((prev) => prev.filter((m) => m.id !== optimisticId))
-          return
-        }
-
-        const currentChatId = await ensureChatExists(uid)
-
-        if (!currentChatId) {
-          setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
-          return
-        }
-
-        const options = {
-          body: {
-            chatId: currentChatId,
-            userId: uid,
-            model: selectedModel,
-            isAuthenticated,
-            systemPrompt: SYSTEM_PROMPT_DEFAULT,
-          },
-        }
-
-        append(
-          {
-            role: "user",
-            content: suggestion,
-          },
-          options
-        )
-        setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
-      } catch (suggestionError) {
-        setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
-        toast({ title: "Failed to send suggestion", status: "error" })
-      } finally {
-        setIsSubmitting(false)
-      }
-    },
-    [
-      ensureChatExists,
-      selectedModel,
-      user,
-      append,
-      checkLimitsAndNotify,
-      isAuthenticated,
-      setMessages,
-    ]
-  )
-
-  const handleReload = useCallback(async () => {
-    const uid = await getOrCreateGuestUserId(user)
-    if (!uid) {
-      return
-    }
-
-    const options = {
-      body: {
-        chatId,
-        userId: uid,
-        model: selectedModel,
-        isAuthenticated,
-        systemPrompt: systemPrompt || SYSTEM_PROMPT_DEFAULT,
-      },
-    }
-
-    reload(options)
-  }, [user, chatId, selectedModel, isAuthenticated, systemPrompt, reload])
+    chatId,
+    reload,
+    setIsSubmitting,
+  })
 
   // Memoize the conversation props to prevent unnecessary rerenders
   const conversationProps = useMemo(
@@ -450,10 +228,6 @@ export function Chat() {
       )}
     >
       <DialogAuth open={hasDialogAuth} setOpen={setHasDialogAuth} />
-
-      <Suspense>
-        <SearchParamsProvider setInput={setInput} />
-      </Suspense>
 
       <AnimatePresence initial={false} mode="popLayout">
         {showOnboarding ? (
