+++ app/components/chat/use-chat-actions.ts
@@ -1,0 +1,278 @@
+import { toast } from "@/components/ui/toast"
+import { getOrCreateGuestUserId } from "@/lib/api"
+import { MESSAGE_MAX_LENGTH, SYSTEM_PROMPT_DEFAULT } from "@/lib/config"
+import { Attachment } from "@/lib/file-handling"
+import type { Message } from "@ai-sdk/react"
+import { useCallback } from "react"
+
+type UseChatActionsProps = {
+  user: any
+  files: File[]
+  createOptimisticAttachments: (files: File[]) => any[]
+  input: string
+  setMessages: (messages: Message[] | ((prev: Message[]) => Message[])) => void
+  setInput: (input: string) => void
+  setFiles: (files: File[]) => void
+  checkLimitsAndNotify: (uid: string) => Promise<boolean>
+  cleanupOptimisticAttachments: (attachments?: any[]) => void
+  ensureChatExists: (uid: string) => Promise<string | null>
+  handleFileUploads: (
+    uid: string,
+    chatId: string
+  ) => Promise<Attachment[] | null>
+  selectedModel: string
+  isAuthenticated: boolean
+  systemPrompt: string
+  enableSearch: boolean
+  handleSubmit: (event?: React.FormEvent, options?: any) => void
+  append: (message: any, options?: any) => void
+  cacheAndAddMessage: (message: Message) => void
+  clearDraft: () => void
+  messages: Message[]
+  bumpChat: (chatId: string) => void
+  chatId: string | null
+  reload: (options?: any) => void
+  setIsSubmitting: (value: boolean) => void
+}
+
+export function useChatActions({
+  user,
+  files,
+  createOptimisticAttachments,
+  input,
+  setMessages,
+  setInput,
+  setFiles,
+  checkLimitsAndNotify,
+  cleanupOptimisticAttachments,
+  ensureChatExists,
+  handleFileUploads,
+  selectedModel,
+  isAuthenticated,
+  systemPrompt,
+  enableSearch,
+  handleSubmit,
+  append,
+  cacheAndAddMessage,
+  clearDraft,
+  messages,
+  bumpChat,
+  chatId,
+  reload,
+  setIsSubmitting,
+}: UseChatActionsProps) {
+  const submit = useCallback(async () => {
+    setIsSubmitting(true)
+
+    const uid = await getOrCreateGuestUserId(user)
+    if (!uid) {
+      setIsSubmitting(false)
+      return
+    }
+
+    const optimisticId = `optimistic-${Date.now().toString()}`
+    const optimisticAttachments =
+      files.length > 0 ? createOptimisticAttachments(files) : []
+
+    const optimisticMessage = {
+      id: optimisticId,
+      content: input,
+      role: "user" as const,
+      createdAt: new Date(),
+      experimental_attachments:
+        optimisticAttachments.length > 0 ? optimisticAttachments : undefined,
+    }
+
+    setMessages((prev) => [...prev, optimisticMessage])
+    setInput("")
+
+    const submittedFiles = [...files]
+    setFiles([])
+
+    try {
+      const allowed = await checkLimitsAndNotify(uid)
+      if (!allowed) {
+        setMessages((prev) => prev.filter((m) => m.id !== optimisticId))
+        cleanupOptimisticAttachments(optimisticMessage.experimental_attachments)
+        return
+      }
+
+      const currentChatId = await ensureChatExists(uid)
+      if (!currentChatId) {
+        setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
+        cleanupOptimisticAttachments(optimisticMessage.experimental_attachments)
+        return
+      }
+
+      if (input.length > MESSAGE_MAX_LENGTH) {
+        toast({
+          title: `The message you submitted was too long, please submit something shorter. (Max ${MESSAGE_MAX_LENGTH} characters)`,
+          status: "error",
+        })
+        setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
+        cleanupOptimisticAttachments(optimisticMessage.experimental_attachments)
+        return
+      }
+
+      let attachments: Attachment[] | null = []
+      if (submittedFiles.length > 0) {
+        attachments = await handleFileUploads(uid, currentChatId)
+        if (attachments === null) {
+          setMessages((prev) => prev.filter((m) => m.id !== optimisticId))
+          cleanupOptimisticAttachments(
+            optimisticMessage.experimental_attachments
+          )
+          return
+        }
+      }
+
+      const options = {
+        body: {
+          chatId: currentChatId,
+          userId: uid,
+          model: selectedModel,
+          isAuthenticated,
+          systemPrompt: systemPrompt || SYSTEM_PROMPT_DEFAULT,
+          enableSearch,
+        },
+        experimental_attachments: attachments || undefined,
+      }
+
+      handleSubmit(undefined, options)
+      setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
+      cleanupOptimisticAttachments(optimisticMessage.experimental_attachments)
+      cacheAndAddMessage(optimisticMessage)
+      clearDraft()
+
+      // Bump existing chats to top (non-blocking, after submit)
+      // If messages.length === 0, this is a new chat that was just created
+      if (messages.length > 0) {
+        bumpChat(currentChatId)
+      }
+    } catch (submitError) {
+      setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
+      cleanupOptimisticAttachments(optimisticMessage.experimental_attachments)
+      toast({ title: "Failed to send message", status: "error" })
+    } finally {
+      setIsSubmitting(false)
+    }
+  }, [
+    user,
+    files,
+    createOptimisticAttachments,
+    input,
+    setMessages,
+    setInput,
+    setFiles,
+    checkLimitsAndNotify,
+    cleanupOptimisticAttachments,
+    ensureChatExists,
+    handleFileUploads,
+    selectedModel,
+    isAuthenticated,
+    systemPrompt,
+    enableSearch,
+    handleSubmit,
+    cacheAndAddMessage,
+    clearDraft,
+    messages.length,
+    bumpChat,
+    setIsSubmitting,
+  ])
+
+  const handleSuggestion = useCallback(
+    async (suggestion: string) => {
+      setIsSubmitting(true)
+      const optimisticId = `optimistic-${Date.now().toString()}`
+      const optimisticMessage = {
+        id: optimisticId,
+        content: suggestion,
+        role: "user" as const,
+        createdAt: new Date(),
+      }
+
+      setMessages((prev) => [...prev, optimisticMessage])
+
+      try {
+        const uid = await getOrCreateGuestUserId(user)
+
+        if (!uid) {
+          setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
+          return
+        }
+
+        const allowed = await checkLimitsAndNotify(uid)
+        if (!allowed) {
+          setMessages((prev) => prev.filter((m) => m.id !== optimisticId))
+          return
+        }
+
+        const currentChatId = await ensureChatExists(uid)
+
+        if (!currentChatId) {
+          setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
+          return
+        }
+
+        const options = {
+          body: {
+            chatId: currentChatId,
+            userId: uid,
+            model: selectedModel,
+            isAuthenticated,
+            systemPrompt: SYSTEM_PROMPT_DEFAULT,
+          },
+        }
+
+        append(
+          {
+            role: "user",
+            content: suggestion,
+          },
+          options
+        )
+        setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
+      } catch (suggestionError) {
+        setMessages((prev) => prev.filter((msg) => msg.id !== optimisticId))
+        toast({ title: "Failed to send suggestion", status: "error" })
+      } finally {
+        setIsSubmitting(false)
+      }
+    },
+    [
+      ensureChatExists,
+      selectedModel,
+      user,
+      append,
+      checkLimitsAndNotify,
+      isAuthenticated,
+      setMessages,
+      setIsSubmitting,
+    ]
+  )
+
+  const handleReload = useCallback(async () => {
+    const uid = await getOrCreateGuestUserId(user)
+    if (!uid) {
+      return
+    }
+
+    const options = {
+      body: {
+        chatId,
+        userId: uid,
+        model: selectedModel,
+        isAuthenticated,
+        systemPrompt: systemPrompt || SYSTEM_PROMPT_DEFAULT,
+      },
+    }
+
+    reload(options)
+  }, [user, chatId, selectedModel, isAuthenticated, systemPrompt, reload])
+
+  return {
+    submit,
+    handleSuggestion,
+    handleReload,
+  }
+}
